refactor(esci2): extract buildDiagnostic + UnsupportedDialectError into diagnostic.ts
